pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-cli-orchestrator"
version = "0.1.2"
description = "Neutral multi-provider CLI orchestrator for agent-style run/review workflows."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "multi-cli-orchestrator maintainers" }
]
keywords = ["cli", "orchestrator", "ai", "agents", "review"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Environment :: Console",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Testing",
]
dependencies = []

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]

[project.urls]
Homepage = "https://github.com/tsk/multi-cli-orchestrator"
Repository = "https://github.com/tsk/multi-cli-orchestrator"
Issues = "https://github.com/tsk/multi-cli-orchestrator/issues"

[project.scripts]
mco = "runtime.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["runtime*"]

[tool.setuptools.package-data]
runtime = ["schemas/*.json"]

